HacksCounter
New admin feature to detect hacking attempts

New database field : tab_uni 'hacks'=>'INT' at the end.

+++ game/pages/admin_debug.php
@@ -9,23 +9,43 @@ function Admin_Debug ()
     global $db_prefix;
     global $GlobalUser;
 
-    if ( method () === "POST" )
+    $message_limit = 50;
+
+    if ( key_exists ('filter', $_REQUEST) ) $filter = $_REQUEST['filter'];
+    else $filter = "";
+
+    if ( method () === "POST" && $filter === "" )
     {
-        $query = "SELECT * FROM ".$db_prefix."debug ORDER BY date DESC, error_id DESC LIMIT 50";
-        $result = dbquery ($query);
-        $rows = dbrows ($result);
-        while ($rows--)
+        if ( $_POST['deletemessages'] === "deleteall" )
         {
-            $msg = dbarray ( $result );
-            if ( $_POST["delmes".$msg['error_id']] === "on" || $_POST['deletemessages'] === "deleteall" )
+            $query = "TRUNCATE TABLE ".$db_prefix."debug;";
+            dbquery ($query);
+        }
+        else
+        {
+            $query = "SELECT * FROM ".$db_prefix."debug ORDER BY date DESC, error_id DESC LIMIT " . $message_limit;
+            $result = dbquery ($query);
+            $rows = dbrows ($result);
+            while ($rows--)
             {
-                $query = "DELETE FROM ".$db_prefix."debug WHERE error_id = " . $msg['error_id'];
-                dbquery ($query);
+                $msg = dbarray ( $result );
+                if ( key_exists ("delmes".$msg['error_id'], $_POST) || $_POST['deletemessages'] === "deleteshown" )
+                {
+                    $query = "DELETE FROM ".$db_prefix."debug WHERE error_id = " . $msg['error_id'];
+                    dbquery ($query);
+                }
             }
         }
     }
 
-    $query = "SELECT * FROM ".$db_prefix."debug ORDER BY date DESC, error_id DESC LIMIT 50";
+    if ( $filter === "" )
+    {
+        $query = "SELECT * FROM ".$db_prefix."debug ORDER BY date DESC, error_id DESC LIMIT " . $message_limit;
+    }
+    else
+    {
+        $query = "SELECT * FROM ".$db_prefix."debug WHERE text LIKE '%".$filter."%' ORDER BY date DESC, error_id DESC LIMIT " . $message_limit;
+    }
     $result = dbquery ($query);
 
 ?>
@@ -52,12 +72,16 @@ function Admin_Debug ()
 
 <tr><td class="b"> </td><td class="b" colspan="3"></td></tr>
 <tr><th colspan="4" style='padding:0px 105px;'></th></tr>
-<tr><th colspan="4">
+<tr>
+<th colspan="4">
 <select name="deletemessages">
 <option value="deletemarked">Удалить выделенные сообщения</option> 
+<option value="deleteshown">Удалить показанные сообщения</option> 
 <option value="deleteall">Удалить все сообщения</option> 
 </select><input type="submit" value="ok" /></th></tr>
 <tr><td colspan="4"><center>     </center></td></tr>
+<tr><th colspan="4">Фильтр отладочных сообщений: <input type=text name="filter" />
+<input type=submit value="Показать"></th></tr>
 </form>
 </table>
